Add volumetric model support
Support reading and training SOM with volumetric models. I created an
InputData class to manage random input from the "dataset", which is just
the vertex positions from the imported model.

The render options and conditions are changed to manage two different
dataset sources: polygonal models and volumentric models. Some
functions, variables, and labels are renamed with the term "model".

A sample pattern is added for future watermarking feature.